make runs_v2 to new run (#490)
* make runs_v2 to new run

* make tables work
