+++ ui/runs/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class RunsConfig(AppConfig):
+class RunsV2Config(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "runs"
