+++ ui/runs/form_mapping.py
@@ -4,10 +4,10 @@ import protzilla.methods.data_analysis as data_analysis
 import protzilla.methods.data_integration as data_integration
 import protzilla.methods.data_preprocessing as data_preprocessing
 import protzilla.methods.importing as importing
-import ui.runs_v2.forms.data_analysis as data_analysis_forms
-import ui.runs_v2.forms.data_integration as data_integration_forms
-import ui.runs_v2.forms.data_preprocessing as data_preprocessing_forms
-import ui.runs_v2.forms.importing as importing_forms
+import ui.runs.forms.data_analysis as data_analysis_forms
+import ui.runs.forms.data_integration as data_integration_forms
+import ui.runs.forms.data_preprocessing as data_preprocessing_forms
+import ui.runs.forms.importing as importing_forms
 from protzilla.run_v2 import Run
 from protzilla.steps import Step
