+++ ui/runs/forms/base.py
@@ -6,7 +6,7 @@ from django.urls import reverse
 
 from protzilla.run_v2 import Run
 from protzilla.utilities import get_file_name_from_upload_path
-from ui.runs_v2.forms.custom_fields import CustomCharField, CustomFileField
+from ui.runs.forms.custom_fields import CustomCharField, CustomFileField
 
 
 class MethodForm(Form):
@@ -44,7 +44,7 @@ class MethodForm(Form):
                 field.widget.attrs.update(
                     {
                         "hx-post": reverse(
-                            "runs_v2:fill_form", kwargs={"run_name": run.run_name}
+                            "runs:fill_form", kwargs={"run_name": run.run_name}
                         ),
                         "hx-target": "#method-form",
                     }
