--- ui/runs/templates/runs/field_checkbox.html
@@ -1,5 +1,0 @@
-{#  params: name, key, default, class #}
-
-<input type="checkbox" id="{{ key }}" name="{{ key }}" value="checked"
-        {% if default %} checked {% endif %} class="{{ class }}"/>
-{% if name %}<label for="{{ key }}">{{ name }}</label>{% endif %}
