--- ui/runs/templates/runs/field_empty.html
@@ -1,2 +1,0 @@
-<label hidden for="{{ key }}">{{ name }}</label>
-<input hidden name="{{ key }}">
