--- ui/runs/templates/runs/field_file.html
@@ -1,4 +1,0 @@
-<label for="{{ key }}" class="mt-2">{{ name }}</label><br>
-<input name="{{ key }}" type="file" id="{{ key }}" class="file-input {{ class }}" hidden>
-<label for="{{ key }}" class="btn btn-grey">Choose file</label>
-<span class="ms-1"><em id="chosen-{{ key }}">No file chosen</em></span>
