--- ui/runs/templates/runs/field_multi_named.html
@@ -1,30 +1,0 @@
-{% if class %}
-<div class="{{ class }} p-3 mb-3 bg-light rounded shadow-sm" id="{{ key }}">
-{% endif %}
-    <label for="{{ key }}" class="fw-bold mb-2">{{ name }}</label>
-    <select name="{{ key }}" id="{{ key }}" class="form-select" {% if disabled %} disabled {% endif %}>
-        {% for option in steps %}
-            <option value="{{ option }}" {% if option == default.0 %} selected {% endif %}>
-                {{ option }}
-            </option>
-        {% endfor %}
-    </select>
-    
-    <!-- Toggle button for Additional Information -->
-    <a href="#{{ key }}-info" class="text-primary mt-2 d-inline-block collapsed" data-bs-toggle="collapse" role="button" aria-expanded="false" aria-controls="{{ key }}-info">
-        View Extracted Values
-    </a>
-
-    <!-- Collapsible Additional Information Content -->
-    <div class="collapse bg-secondary bg-opacity-10 rounded p-3 mt-1 shadow-sm" id="{{ key }}-info">
-        <p>The following values are being extracted from the chosen dictionary:</p>
-        <hr>
-        <ul class="list-group list-group-flush">
-            {% for item in mapping_keys %}
-                <li class="list-group-item bg-transparent py-2 px-3 border-0 mb-1">{{ item }}</li>
-            {% endfor %}
-        </ul>
-    </div>
-{% if class %}
-</div>
-{% endif %}
