--- ui/runs/templates/runs/field_select.html
@@ -1,10 +1,0 @@
-{#  params: name, categories, key, disabled, multiple, default#}
-{% if name %}<label for="{{ key }}" >{{ name }}</label>{% endif %}
-<select name="{{ key }}" id="{{ key }}"
-        class="form-select {{ class }}"
-        {% if disabled %} disabled {% endif %}
-        {% if multiple %} multiple {% endif %}>
-    {% for option in categories %}
-        <option value="{{ option }}" {% if option in default %} selected="selected" {% endif %} >{{ option }}</option>
-    {% endfor %}
-</select>
