--- ui/runs/templates/runs/field_select_dynamic.html
@@ -1,14 +1,0 @@
-{#  params: name, categories, key, disabled, multiple, dynamic_fields, default#}
-{% csrf_token %}
-{%if name %}<label for="{{ key }}">{{ name }}</label>{% endif %}
-<select name="{{ key }}" id="{{ key }}"
-        class="form-select dynamic_parameter_trigger"
-        {% if disabled %} disabled {% endif %}
-        {% if multiple %} multiple {% endif %}>
-    {% for option in categories %}
-        <option value="{{ option }}" {% if option in default%} selected="selected" {% endif %} >{{ option }}</option>
-    {% endfor %}
-</select>
-<div id="dynamic_fields_{{ key }}">
-{% include "runs/fields.html" with fields=dynamic_fields %}
-</div>
