--- ui/runs/templates/runs/field_text.html
@@ -1,11 +1,0 @@
-{# params: name, key, default, disabled, form #}
-
-{# key: the name of the input #}
-{# default: the default value of the input #}
-{# disabled: if the input is disabled #}
-{# form: the name of the form #}
-
-<label for="{{ key }}">{{ name }}</label>
-<input name="{{ key }}" type="text" id="{{ key }}" value="{{ default }}"
-    class="form-control {{ class }}" {% if disabled %} disabled {% endif %}
-    placeholder="{{ placeholder }}">
