--- ui/runs/templates/runs/fields.html
@@ -1,5 +1,0 @@
-{% for field in fields %}
-    <div class="mb-2">
-        {{ field }}
-    </div>
-{% endfor %}
