--- ui/runs/templates/runs/fields_classification_methods.html
@@ -1,54 +1,0 @@
-{% load id_tags %}
-{% if history %}
-    {% for field in fields %}
-        <div class="mb-2">
-            {% if field|has_element_id:"train_test_split" %}
-                <hr>
-                <em>Test-Train Ratio</em>
-                <br>
-                {{ field }}
-            {% elif field|has_element_id:"validation_strategy" %}
-                <hr>
-                <em>Validation Strategy</em>
-                <br>
-                {{ field }}
-            {% elif field|has_element_id:"model_selection" %}
-                <hr>
-                <em>Hyperparameter fine-tuning</em>
-                <br>
-                {{ field }}
-            {% else %}
-                {{ field }}
-            {% endif %}
-        </div>
-    {% endfor %}
-{% else %}
-    {% for field in fields %}
-        <div class="mb-2">
-            {% if field|has_element_id:"train_test_split" %}
-                <hr>
-                <strong>Test-Train Ratio</strong>
-                <p><em> For choosing the size of the test data set you can either enter the
-                    absolute number of test samples or a number between 0.0 and 1.0 to
-                    represent the percentage of test samples</em></p>
-                {{ field }}
-            {% elif field|has_element_id:"validation_strategy" %}
-                <hr>
-                <strong>Validation Strategy</strong>
-                <br>
-                {{ field }}
-            {% elif field|has_element_id:"model_selection" %}
-                <hr>
-                <strong>Hyperparameter fine-tuning</strong>
-                <p><em>To find the best hyperparameters for your model choose your preferred
-                    grid search method and select multiple values for a specific parameter.
-                    If the parameter is categorical, use the multiselect function to choose
-                    multiple options. For numerical inputs, separate the different numbers
-                    by using a pipe “|”, e.g. 1|2|0.12 </em></p>
-                {{ field }}
-            {% else %}
-                {{ field }}
-            {% endif %}
-        </div>
-    {% endfor %}
-{% endif %}
