--- ui/runs/templates/runs/form_add_name.html
@@ -1,9 +1,0 @@
-{# params: run_name index #}
-<form action="{% url 'runs:add_name' run_name %}" method="post" class="col">
-    {% csrf_token %}
-    <div class="input-group">
-        <input type="text" name="name" class="form-control" placeholder="name the output of this step">
-        <input type="hidden" name="index" value="{{ index }}">
-        <input type="submit" value="add" class="btn btn-grey">
-    </div>
-</form>
