--- ui/runs/templates/runs/form_buttons.html
@@ -1,217 +1,0 @@
-<script>
-    // This method takes care of the plot and calculation buttons. The goal is to prevent users from calculating twice
-    // for the same calculation parameters and from plotting when the selected calculation parameters don't match the
-    // ones that results are present for so there is no confusion about what calculation results was plotted for
-
-    // Control flow: all relevant information is gathered via a GET-request from the backend (`get_all_parameters`) and,
-    // when needed via `get_results_exist`. To ensure the relevant information is present when deciding whether to
-    // enable or disable the buttons, `get_all_parameters` takes a callback function that triggers the .change()
-    // listeners after a successful GET. The .change() listeners are also triggered every time the calculation method
-    // is changed (see ./dynamic_methods::$("#chosen_method").change(...)).
-    // The plot button is always updated after the calculation button.
-
-    // Admittedly, this is not the most robust construct, and it took way longer to figure out all the edge cases than
-    // I would like to admit. git blame can be quite useful ;)
-
-    let current_parameters;
-    let current_plot_parameters;
-    let calc_params_match;
-    let plotted_for_parameters;
-    let calc_form_data;
-    let calculated_method;
-
-    function get_all_parameters(callback) {
-        $.ajax({
-            url: "{% url 'runs:all_button_parameters' run_name %}",
-            type: "GET",
-            success: function (response) {
-                current_parameters = response["current_parameters"];
-                current_plot_parameters = response["current_plot_parameters"];
-                plotted_for_parameters = response["plotted_for_parameters"];
-                calculated_method = response["chosen_method"];
-                callback();
-            }
-        }).fail(function (e) {
-            console.log('Error', e);
-        });
-    }
-
-    async function get_results_exist(callback) {
-        $.ajax({
-            url: "{% url 'runs:results_exist' run_name %}",
-            type: "GET",
-            success: function (response) {
-                results_exist = response["results_exist"];
-                callback();
-            }
-        }).fail(function (e) {
-            console.log('Error', e);
-        });
-    }
-
-    function get_clean_form_data(form) {
-        let formDataArray = form.serializeArray();
-        let formDataDict = {};
-        for (let formField of formDataArray) {
-            formDataDict[formField.name] = formField.value;
-        }
-        delete formDataDict["csrfmiddlewaretoken"];
-        return formDataDict;
-    }
-
-    function disable_calc_button(bool) {
-        $('#calculate_parameters_submit').prop('disabled', bool);
-        results_exist = bool;
-        calc_params_match = bool;
-    }
-
-    function disable_plot_button(bool) {
-        plot_button = $('#plot_parameters_submit');
-        //When plot fields exist the Plot button is implemented as an input otherwise as a link,
-        //thus they have to be deactivated differently.
-        if (plot_fields_exist()) {
-            plot_button.prop('disabled', bool);
-        }
-        else {
-            if (bool) {
-                plot_button.prop('href', '');
-                plot_button.prop('style', 'background-color:#EFF3F7');
-                plot_button.prop('style', 'color:#6E7073');
-            }
-            else {
-                plot_button.prop('href', '{% url 'runs:plot' run_name %}');
-            }
-        }
-    }
-
-    function disable_plot_fields(bool) {
-        $('#plotfields').prop('disabled', bool);
-    }
-
-    function isEmptyDict(obj) {
-        return Object.keys(obj).length === 0;
-    }
-
-    function plot_fields_exist() {
-        return document.getElementById('plot_form') !== null;
-    }
-
-    function trigger_plot_change() {
-        get_results_exist(function () {
-            $('#plot_form').trigger('change');
-        })
-    }
-
-    function plotted_current_match() {
-        if (isEmptyDict(plotted_for_parameters)) {
-            return false;
-        }
-        return Object.keys(current_parameters).every(key =>
-            plotted_for_parameters.hasOwnProperty(key) && plotted_for_parameters[key] == current_parameters[key]
-        );
-    }
-
-    function equal_calc_parameters(form) {
-        if (isEmptyDict(current_parameters)) {
-            return false;
-        }
-
-        let form_data = get_clean_form_data(form);
-        calc_form_data = form_data;
-        return Object.keys(current_parameters).every(key =>
-            form_data.hasOwnProperty(key) && form_data[key] == current_parameters[key]
-        );
-    }
-
-
-    function showCalculationInProgressModal() {
-        $('#calculationInProgressModal').modal('show');
-    }
-
-    function hideCalculationInProgressModal() {
-        $('#calculationInProgressModal').modal('hide');
-    }
-
-
-    function equal_plot_parameters(form) {
-        if (isEmptyDict(current_plot_parameters)) {
-            return false;
-        }
-
-        let form_data = get_clean_form_data(form);
-        return Object.keys(current_plot_parameters).every(key =>
-            form_data.hasOwnProperty(key) && form_data[key] == current_plot_parameters[key]
-        );
-    }
-
-    $(document).ready(function () {
-        disable_calc_button(false);
-        calculationInProgress = false;
-        $('#backButton').click(function () {
-            if (calculationInProgress) {
-                event.preventDefault();
-                showCalculationInProgressModal();
-                $('#cancel').click(function () {
-                    hideCalculationInProgressModal();
-            });
-    }
-    });
-        if($('#plot_form').find('input').length) {
-            disable_plot_button(true);
-            disable_plot_fields(true);
-        }
-
-        $('#calc_form').change(function () {
-            calculationInProgress = true;
-            const selected_method = $('#chosen_method').val();
-            calc_form_data = get_clean_form_data($(this));
-            if (equal_calc_parameters($(this)) && (calculated_method === selected_method)) {
-                disable_calc_button(true);
-                disable_plot_fields(false);
-            } else {
-                disable_calc_button(false);
-                disable_plot_fields(true);
-            }
-            if (!plot_fields_exist()) {
-                if (equal_calc_parameters($(this)) && (calculated_method === selected_method)) {
-                    disable_plot_button(false);
-                } else {
-                    disable_plot_button(true);
-                }
-            }
-            calculationInProgress = false;
-            $('#plot_form').trigger("change");
-            
-        });
-
-        $('#plot_form').change(function () {
-            const selected_method = $('#chosen_method').val();
-
-            if (plotted_current_match() && equal_plot_parameters($(this))) {
-                disable_plot_button(true);
-            }
-
-            if (!calc_form_data || selected_method !== calculated_method) {
-                disable_plot_button(true)
-                return;
-            }
-
-            if (!plotted_current_match() && results_exist) {
-                disable_plot_button(false);
-                return;
-            }
-
-            if (equal_plot_parameters($(this)) || !calc_params_match) {
-                disable_plot_button(true);
-            }
-            else {
-                disable_plot_button(false);
-            }
-        });
-
-        get_all_parameters(function () {
-            $('#calc_form').trigger('change');
-            trigger_plot_change();
-        });
-    });
-</script>
