+++ ui/runs/templates/runs/method_dropdown.html
@@ -1,5 +1,5 @@
 <label for="{{ key }}">{{ name }}</label>
-<select name="{{ key }}" class="form-select" hx-post="{% url 'runs_v2:change_method' run_name %}" hx-target="body" hx-swap="outerHTML">
+<select name="{{ key }}" class="form-select" hx-post="{% url 'runs:change_method' run_name %}" hx-target="body" hx-swap="outerHTML">
     {% for option, label in categories %}
         <option value="{{ option }}" {% if option == default %} selected="selected" {% endif %}
                 label="{{ label }}">{{ option }}</option>
