--- ui/runs/templates/runs/plot_width.html
@@ -1,17 +1,0 @@
-<script>
-    // resize the plotly plots when they are displayed
-    // to avoid errors with the width
-    $(document).ready(function () {
-        let plotsAreInHistory = false;
-        {% for step in displayed_history %}
-            {% if step.plots %}
-            // at least one plot is being displayed
-            plotsAreInHistory = true;
-            {% endif %}
-        {% endfor %}
-
-        if (plotsAreInHistory) {
-            window.dispatchEvent(new Event('resize'));
-        }
-    });
-</script>
