+++ ui/runs/templates/runs/sidebar_section.html
@@ -13,26 +13,22 @@
 <div>
     {% for step in section.steps %}
         <div class="wrapper {% if step.selected %}selected{% endif %} {% if step.finished %}finished{% endif %} d-flex align-items-center">
-            {% if not step.finished %}
-                <p class="me-1">{{ step.name }}: <i>{{ step.method_name }}</i></p>
-            {% else %}
-                <form action="{% url 'runs:navigate' run_name %}" method="post">
-                    {% csrf_token %}
-                    <input type="hidden" name="section_name" value="{{ section.id }}">
-                    <input type="hidden" name="index" value="{{ step.index }}">
-                    <button class="me-1" style="border: none; background: none; text-align: left">
-                        <p class="me-1">{{ step.name }}: <i>{{ step.method_name }}</i></p>
-                    </button>
-                </form>
-            {% endif %}
+            <form action="{% url 'runs:navigate' run_name %}" method="post">
+                {% csrf_token %}
+                <input type="hidden" name="section_name" value="{{ section.id }}">
+                <input type="hidden" name="index" value="{{ step.index }}">
+                <button class="me-1" style="border: none; background: none; text-align: left">
+                    <p class="me-1">{{ step.name }}: <i>{{ step.method_name }}</i></p>
+                </button>
+            </form>
 
             {% if not step.selected and not step.finished %}
                 {# button to remove step from workflow #}
                 <form action="{% url 'runs:delete_step' run_name %}" method="post">
                     {% csrf_token %}
-                    <input type="hidden" name="section_name" value="{{ section.id }}">
                     <input type="hidden" name="index" value="{{ step.index }}">
-                    <button class="removeBtn btn d-flex justify-content-center align-items-end mt-1">                    
+                    <input type="hidden" name="section" value="{{ step.section }}">
+                    <button class="removeBtn btn d-flex justify-content-center align-items-end mt-1">
                         <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="grey" class="bi bi-x removeIcon">
                         <path d="M4.646 4.646a.5.5 0 0 1 .708 0L8 7.293l2.646-2.647a.5.5 0 0 1 .708.708L8.707 8l2.647 2.646a.5.5 0 0 1-.708.708L8 8.707l-2.646 2.647a.5.5 0 0 1-.708-.708L7.293 8 4.646 5.354a.5.5 0 0 1 0-.708z"/>
                     </svg></button>
