--- ui/runs/templatetags/id_tags.py
@@ -1,13 +1,0 @@
-from bs4 import BeautifulSoup
-from django import template
-
-register = template.Library()
-
-
-@register.filter
-def has_element_id(field, id_name):
-    soup = BeautifulSoup(field, "html.parser")
-    element = soup.find(id=id_name)
-    if element and element.get("id") == id_name:
-        return True
-    return False
