+++ ui/runs/views.py
@@ -1,6 +1,4 @@
-from __future__ import annotations
-
-import sys
+import io
 import tempfile
 import traceback
 import zipfile
@@ -11,63 +9,44 @@ import numpy as np
 import pandas as pd
 from django.http import (
     FileResponse,
+    HttpRequest,
+    HttpResponse,
     HttpResponseBadRequest,
     HttpResponseRedirect,
     JsonResponse,
 )
 from django.shortcuts import render
-from django.template.loader import render_to_string
 from django.urls import reverse
-from main.settings import BASE_DIR
 
-sys.path.append(f"{BASE_DIR}/..")
-
-from protzilla.constants.protzilla_logging import logger
-from protzilla.data_integration.database_query import uniprot_columns
-from protzilla.run_helper import get_parameters, log_messages
-from protzilla.run_v2 import Run
-from protzilla.utilities import (
-    clean_uniprot_id,
+from protzilla.run_helper import log_messages
+from protzilla.run_v2 import Run, get_available_run_names
+from protzilla.stepfactory import StepFactory
+from protzilla.steps import Step
+from protzilla.utilities.utilities import (
+    check_is_path,
+    format_trace,
     get_memory_usage,
     name_to_title,
-    unique_justseen,
 )
+from protzilla.workflow import get_available_workflow_names
 from ui.runs.fields import (
-    make_current_fields,
     make_displayed_history,
-    make_dynamic_fields,
     make_method_dropdown,
     make_name_field,
-    make_parameter_input,
-    make_plot_fields,
     make_sidebar,
 )
-from ui.runs.views_helper import display_messages, parameters_from_post
+from ui.runs.views_helper import display_message, display_messages, parameters_from_post
+
+from .form_mapping import (
+    get_empty_plot_form_by_method,
+    get_filled_form_by_method,
+    get_filled_form_by_request,
+)
 
 active_runs: dict[str, Run] = {}
 
 
-def index(request):
-    """
-    Renders the main index page of the PROTzilla application.
-
-    :param request: the request object
-    :type request: HttpRequest
-
-    :return: the rendered index page
-    :rtype: HttpResponse
-    """
-    return render(
-        request,
-        "runs/index.html",
-        context={
-            "available_workflows": WorkflowManager.available_workflows(),
-            "available_runs": Run.available_runs(),
-        },
-    )
-
-
-def detail(request, run_name):
+def detail(request: HttpRequest, run_name: str):
     """
     Renders the details page of a specific run.
     For rendering a context dict is created that contains all the dynamic information
@@ -83,15 +62,30 @@ def detail(request, run_name):
     :rtype: HttpResponse
     """
     if run_name not in active_runs:
-        active_runs[run_name] = Run.run_read(run_name)
-    run = active_runs[run_name]
-    # TODO we do not need current_run_location anymore in that style
-    # section, step, method = run.steps.current_step_long()
-    last_step = run.steps.current_step_index == len(run.steps.all_steps) - 1
+        active_runs[run_name] = Run(run_name)
+    run: Run = active_runs[run_name]
 
-    log_messages(run.steps.current_step.messages)
-    display_messages(run.steps.current_step.messages, request)
-    run.steps.current_step.messages = []
+    # section, step, method = run.current_run_location()
+    # end_of_run = not step
+
+    if request.POST:
+        method_form = get_filled_form_by_request(
+            request, run
+        )  # TODO maybe not do this as it is done after the calculation
+        if method_form.is_valid():
+            method_form.submit(run)
+        plot_form = get_empty_plot_form_by_method(run.current_step, run)
+        # in case the fill_form now would change it
+        method_form.fill_form(run)
+    else:
+        method_form = get_filled_form_by_method(run.current_step, run)
+        plot_form = get_empty_plot_form_by_method(run.current_step, run)
+
+    description = run.current_step.method_description
+
+    log_messages(run.current_step.messages)
+    display_messages(run.current_messages, request)
+    run.current_messages.clear()
 
     current_plots = []
     for plot in run.current_plots:
@@ -114,15 +108,17 @@ def detail(request, run_name):
         else:
             current_plots.append(plot.to_html(include_plotlyjs=False, full_html=False))
 
-    show_table = not run.current_outputs.is_empty and any(
-        isinstance(v, pd.DataFrame) for _, v in run.current_outputs
+    show_table = (
+        not run.current_outputs.is_empty
+        and any(isinstance(v, pd.DataFrame) for _, v in run.current_outputs)
+        or any(check_is_path(v) for _, v in run.current_outputs)
     )
 
     show_protein_graph = (
         run.current_outputs
         and "graph_path" in run.current_outputs
         and run.current_outputs["graph_path"] is not None
-        and Path(run.current_out["graph_path"]).exists()
+        and Path(run.current_outputs["graph_path"]).exists()
     )
 
     return render(
@@ -130,283 +126,61 @@ def detail(request, run_name):
         "runs/details.html",
         context=dict(
             run_name=run_name,
-            section=run.steps.current_step.section,
-            step=run.steps.current_step.step,
-            display_name=f"{name_to_title(run.current_step.name)}",
-            displayed_history=make_displayed_history(run),  # TODO
-            method_dropdown=make_method_dropdown(run, section, step, method),  # TODO
-            fields=make_current_fields(run, section, step, method),
-            plot_fields=make_plot_fields(run, section, step, method),  # TODO
+            section=run.current_step.section,
+            step=run.current_step,
+            display_name=f"{name_to_title(run.current_step.operation)} - {run.current_step.display_name}",
+            displayed_history=make_displayed_history(
+                run
+            ),  # TODO: make NewRun compatible
+            method_dropdown=make_method_dropdown(
+                run.run_name,
+                run.current_step.section,
+                run.current_step.operation,
+                type(run.current_step).__name__,
+            ),
             name_field=make_name_field(
-                run.current_step.finished, run, end_of_run
-            ),  # TODo
+                run.current_step.finished, run, False
+            ),  # TODO end_of_run
             current_plots=current_plots,
             results_exist=run.current_step.finished,
             show_back=run.steps.current_step_index > 0,
-            show_plot_button=not run.current_outputs.is_empty,
-            sidebar=make_sidebar(request, run, run_name),  # TODO
-            last_step=last_step,
-            end_of_run=end_of_run,
+            show_plot_button=run.current_step.finished,
+            # TODO include plot exists and plot parameters match current plot or remove this and replace with results exist
+            sidebar=make_sidebar(request, run),
+            last_step=run.steps.current_step_index == len(run.steps.all_steps) - 1,
+            end_of_run=False,  # TODO?
             show_table=show_table,
             used_memory=get_memory_usage(),
             show_protein_graph=show_protein_graph,
             description=description,
+            method_form=method_form,
+            is_form_dynamic=method_form.is_dynamic,
+            plot_form=plot_form,
         ),
     )
 
 
-def change_method(request, run_name):
+def index(request: HttpRequest, index_error: bool = False):
     """
-    Changes the method during a step of a run.
-    This is called when the user selects a new method in the first dropdown of a step.
+    Renders the main index page of the PROTzilla application.
 
     :param request: the request object
     :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
 
-    :return: a JSON response object containing the new fields for the selected method
-    :rtype: JsonResponse
+    :return: the rendered index page
+    :rtype: HttpResponse
     """
-    # TODO 92 extract into a separate method like try_reactivate_run
-    try:
-        if run_name not in active_runs:
-            active_runs[run_name] = Run.continue_existing(run_name)
-        run = active_runs[run_name]
-    except FileNotFoundError:
-        traceback.print_exc()
-        response = JsonResponse({"error": f"Run '{run_name}' was not found"})
-        response.status_code = 404  # not found
-        return response
-
-    run.method = request.POST["method"]
-    section, step, _ = run.current_workflow_location()
-    run.update_workflow_config([], update_params=False)
-
-    current_fields = make_current_fields(run, run.section, run.step, run.method)
-    plot_fields = make_plot_fields(run, run.section, run.step, run.method)
-    return JsonResponse(
-        dict(
-            parameters=render_to_string(
-                "runs/fields.html",
-                context=dict(fields=current_fields),
-            ),
-            plot_parameters=render_to_string(
-                "runs/fields.html",
-                context=dict(fields=plot_fields),
-            ),
-        ),
-        safe=False,
+    return render(
+        request,
+        "runs/index.html",
+        context={
+            "available_workflows": get_available_workflow_names(),
+            "available_runs": get_available_run_names(),
+        },
     )
 
 
-def change_dynamic_fields(request, run_name):
-    """
-    Renders fields that depend on the value of another field e.g. a dropdown, the value
-    being the dynamic_trigger_value below. The field is specified by its key and part of
-    the request.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: a JSON response object containing the new fields depending on the value of
-        the dynamic trigger
-    :rtype: JsonResponse
-    """
-
-    try:
-        if run_name not in active_runs:
-            active_runs[run_name] = Run.continue_existing(run_name)
-        run = active_runs[run_name]
-    except FileNotFoundError:
-        traceback.print_exc()
-        response = JsonResponse({"error": f"Run '{run_name}' was not found"})
-        response.status_code = 404  # not found
-        return response
-
-    dynamic_trigger_value = request.POST["selected_input"]
-    dynamic_trigger_key = request.POST["key"]
-    all_parameters_dict = get_parameters(run, run.section, run.step, run.method)
-    dynamic_trigger_param_dict = all_parameters_dict[dynamic_trigger_key]
-    dynamic_fields = make_dynamic_fields(
-        dynamic_trigger_param_dict, dynamic_trigger_value, all_parameters_dict, False
-    )
-    parameters = render_to_string(
-        "runs/fields.html",
-        context=dict(fields=dynamic_fields),
-    )
-    return JsonResponse(dict(parameters=parameters), safe=False)
-
-
-def change_field(request, run_name):
-    """
-    Changes the value of one or multiple fields during a method of a run depending on a
-    selected value in another field. The field that triggers this method is identified by
-    the post_id variable.
-    In contrast to change_dynamic_fields, this method changes the value of the field itself
-    instead of rendering new fields.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: a JSON response object containing the updated fields depending on the value of the
-        dynamic trigger field
-    :rtype: JsonResponse
-    """
-    try:
-        if run_name not in active_runs:
-            active_runs[run_name] = Run.continue_existing(run_name)
-        run = active_runs[run_name]
-    except FileNotFoundError as e:
-        print(str(e))
-        response = JsonResponse({"error": f"Run '{run_name}' was not found"})
-        response.status_code = 404  # not found
-        return response
-
-    selected = request.POST.getlist("selected[]")
-    post_id = request.POST["id"]
-    if len(selected) > 1:
-        # remove last 4 characters from post_id to get the original id
-        # because multiple selected items are in id_div
-        post_id = post_id[:-4]
-
-    parameters = run.workflow_meta[run.section][run.step][run.method]["parameters"]
-    if "fill_dynamic" in parameters[post_id]:
-        fields_to_fill = parameters[post_id]["fill_dynamic"]
-    else:
-        fields_to_fill = [
-            k for k in parameters.keys() if parameters[k]["type"] == "named_output_v2"
-        ]
-
-    fields = {}
-    for key in fields_to_fill:
-        param_dict = parameters[key]
-
-        if "fill" in param_dict:
-            if param_dict["fill"] == "metadata_column_data":
-                param_dict["categories"] = run.metadata[selected[0]].unique().tolist()
-            elif param_dict["fill"] == "uniprot_fields":
-                param_dict["categories"] = uniprot_columns(selected[0]) + ["Links"]
-            elif param_dict["fill"] == "protein_ids":
-                if len(selected) == 1:
-                    named_output = run.current_out_sources[selected[0]]
-                    output_item = selected[0]
-                else:
-                    named_output = selected[0]
-                    output_item = selected[1]
-                # KeyError is expected when named_output triggers the fill
-                try:
-                    protein_iterable = run.history.output_of_named_step(
-                        named_output, output_item
-                    )
-                except KeyError:
-                    protein_iterable = None
-                if isinstance(protein_iterable, pd.DataFrame):
-                    param_dict["categories"] = protein_iterable["Protein ID"].unique()
-                elif isinstance(protein_iterable, pd.Series):
-                    param_dict["categories"] = protein_iterable.unique()
-                elif isinstance(protein_iterable, list):
-                    param_dict["categories"] = protein_iterable
-                else:
-                    param_dict["categories"] = []
-                    logger.warning(
-                        f"Warning: expected protein_iterable to be a DataFrame, Series or list, but got {type(protein_iterable)}. Proceeding with empty list."
-                    )
-
-            elif param_dict["fill"] == "protein_df_columns":
-                named_output = selected[0]
-                output_item = selected[1]
-                # KeyError is expected when named_output triggers the fill
-                try:
-                    protein_iterable = run.history.output_of_named_step(
-                        named_output, output_item
-                    )
-                except KeyError:
-                    protein_iterable = None
-                if isinstance(protein_iterable, pd.DataFrame):
-                    categories = []
-                    for column in protein_iterable.columns:
-                        if column not in ["Protein ID", "Sample"]:
-                            categories.append(column)
-                    param_dict["categories"] = categories
-                elif isinstance(protein_iterable, pd.Series):
-                    param_dict["categories"] = protein_iterable.index
-                elif isinstance(protein_iterable, dict):
-                    param_dict["categories"] = protein_iterable.keys()
-                else:
-                    param_dict["categories"] = []
-                    logger.warning(
-                        f"Warning: expected protein_iterable to be a DataFrame or Series, but got {type(protein_iterable)}. Proceeding with empty list."
-                    )
-
-            elif param_dict["fill"] == "enrichment_categories":
-                named_output = selected[0]
-                output_item = selected[1]
-
-                # TODO: this is a bit hacky, but it works for now
-                # should be refactored when we rework the named input handling
-                # KeyError is expected here because named_output trigger change_field
-                # twice to make sure that the categories are updated after the named_output has updated
-                try:
-                    protein_iterable = run.history.output_of_named_step(
-                        named_output, output_item
-                    )
-                except KeyError:
-                    protein_iterable = None
-
-                if (
-                    not isinstance(protein_iterable, pd.DataFrame)
-                    or not "Gene_set" in protein_iterable.columns
-                ):
-                    param_dict["categories"] = []
-                else:
-                    param_dict["categories"] = (
-                        protein_iterable["Gene_set"].unique().tolist()
-                    )
-            elif param_dict["fill"] == "gsea_enrichment_categories":
-                named_output = selected[0]
-                output_item = selected[1]
-
-                try:
-                    protein_iterable = run.history.output_of_named_step(
-                        named_output, output_item
-                    )
-                except KeyError:
-                    protein_iterable = None
-
-                if (
-                    not isinstance(protein_iterable, pd.DataFrame)
-                    or not "NES" in protein_iterable.columns
-                ):
-                    param_dict["categories"] = []
-                else:
-                    # gene_set_libraries are all prefixes for Term column in gsea output
-                    # if no prefix is found, a single gmt file was used
-                    gene_set_libraries = set()
-                    for term in protein_iterable["Term"].unique():
-                        if "__" in term:
-                            gene_set_lib = term.split("__")[0]
-                            gene_set_libraries.add(gene_set_lib)
-                        else:
-                            gene_set_libraries.add("all")
-                    param_dict["categories"] = list(gene_set_libraries)
-
-        elif "select_from" in param_dict:
-            param_dict["outputs"] = run.history.output_keys_of_named_step(selected[0])
-
-            for out_key in param_dict["outputs"]:
-                run.current_out_sources[out_key] = selected[0]
-
-        fields[key] = make_parameter_input(key, param_dict, parameters, disabled=False)
-
-    return JsonResponse(fields, safe=False)
-
-
-def create(request):
+def create(request: HttpRequest):
     """
     Creates a new run. The user is then redirected to the detail page of the run.
 
@@ -417,16 +191,29 @@ def create(request):
     :rtype: HttpResponse
     """
     run_name = request.POST["run_name"]
-    run = Run.create(
-        run_name,
-        request.POST["workflow_config_name"],
-        df_mode=request.POST["df_mode"],
-    )
+    try:
+        run = Run(
+            run_name,
+            request.POST["workflow_config_name"],
+            df_mode=request.POST["df_mode"],
+        )
+    except Exception as e:
+        display_message(
+            {
+                "level": 40,
+                "msg": "Something went wrong creating a new run.",
+                "trace": format_trace(traceback.format_exception(e)),
+            },
+            request,
+        )
+        traceback.print_exc()
+        return HttpResponseRedirect(reverse("runs:index"))
+
     active_runs[run_name] = run
     return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
 
 
-def continue_(request):
+def continue_(request: HttpRequest):
     """
     Continues an existing run. The user is redirected to the detail page of the run and
     can resume working on the run.
@@ -438,8 +225,7 @@ def continue_(request):
     :rtype: HttpResponse
     """
     run_name = request.POST["run_name"]
-
-    active_runs[run_name] = Run.run_read(run_name)
+    active_runs[run_name] = Run(run_name)
 
     return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
 
@@ -457,8 +243,10 @@ def next_(request, run_name):
     :rtype: HttpResponse
     """
     run = active_runs[run_name]
-
-    run.next_step(request.POST["name"])
+    name = request.POST.get("name", None)
+    if name:
+        run.steps.name_current_step_instance(name)
+    run.step_next()
 
     return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
 
@@ -476,115 +264,7 @@ def back(request, run_name):
     :rtype: HttpResponse
     """
     run = active_runs[run_name]
-    run.back_step()
-    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
-
-
-def navigate(request, run_name):
-    """
-    Navigates to a specific step/method of the run. The user is then redirected to the
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-    """
-    run = active_runs[run_name]
-    run.navigate(request.POST["section_name"], int(request.POST["index"]))
-    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
-
-
-def add(request, run_name):
-    """
-    Adds a new method to the run. The method is added as the next step.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: the rendered detail page of the run, new method visible in sidebar
-    :rtype: HttpResponse
-    """
-    run = active_runs[run_name]
-
-    post = dict(request.POST)
-    del post["csrfmiddlewaretoken"]
-    step = post["step"][0]
-    section = post["section_name"][0]
-    method = post["method"][0]
-
-    run.insert_at_next_position(step, section, method)
-    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
-
-
-def delete_step(request, run_name):
-    """
-    Deletes a step/method from the run.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: the rendered detail page of the run, deleted method no longer visible in sidebar
-    :rtype: HttpResponse
-    """
-    run = active_runs[run_name]
-
-    post = dict(request.POST)
-    del post["csrfmiddlewaretoken"]
-    index = int(post["index"][0])
-    section = post["section_name"][0]
-    run.delete_step(section, index)
-    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
-
-
-def export_workflow(request, run_name):
-    """
-    Exports the workflow of the run as a JSON file so that it can be reused and shared.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: the rendered detail page of the run
-    :rtype: HttpResponse
-    """
-    run = active_runs[run_name]
-    post = dict(request.POST)
-    del post["csrfmiddlewaretoken"]
-    print("run_name")
-    print(run_name)
-    name = post["name"][0]
-    run.export_workflow(name)
-
-    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
-
-
-def calculate(request, run_name):
-    """
-    Performs the current methods calculation during the run. Django messages are used to
-    display additional information, warnings and errors to the user.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: the rendered detail page of the run
-    :rtype: HttpResponse
-    """
-    run = active_runs[run_name]
-    parameters = parameters_from_post(request.POST)
-    del parameters["chosen_method"]
-
-    for k, v in dict(request.FILES).items():
-        # assumption: only one file uploaded
-        parameters[k] = v[0].temporary_file_path()
-    run.perform_current_calculation_step(parameters)
-
+    run.step_previous()
     return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
 
 
@@ -604,158 +284,37 @@ def plot(request, run_name):
     :rtype: HttpResponse
     """
     run = active_runs[run_name]
-    section, step, method = run.current_run_location()
     parameters = parameters_from_post(request.POST)
 
-    if run.step == "plot":
+    if run.current_step.display_name == "plot":
         del parameters["chosen_method"]
-
-    run.create_plot_from_current_location(parameters)
-
-    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
-
-
-def add_name(request, run_name):
-    """
-    Adds a name to the results of a calculated method of the run. The name can be used
-    to identify the result and use them later.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: the rendered detail page of the run
-    :rtype: HttpResponse
-    """
-    run = active_runs[run_name]
-    run.name_step(int(request.POST["index"]), request.POST["name"])
-    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
-
-
-def results_exist_json(request, run_name):
-    """
-    Checks if the results of the run exist. This is used to determine if the Next button
-    should be enabled or not.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: a JSON response with a boolean value
-    :rtype: JsonResponse
-    """
-    run = active_runs[run_name]
-    return JsonResponse(dict(results_exist=run.current_step.finished))
-
-
-def all_button_parameters(request, run_name):
-    """
-    Returns all parameters that are needed to render the buttons as enabled or disabled
-    in the run detail page.
-    See ui/runs/templates/runs/form_buttons.html for detailed documentation.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: a JSON response with the parameters
-    :rtype: JsonResponse
-    """
-    run = active_runs[run_name]
-    d = dict()
-    d["current_plot_parameters"] = run.current_plot_parameters.get(run.method, {})
-    d["plotted_for_parameters"] = (
-        run.plotted_for_parameters if run.plotted_for_parameters is not None else dict()
-    )
-
-    if run.current_parameters is None or not run.current_step.finished:
-        d["current_parameters"] = dict()
-        d["chosen_method"] = dict()
+        run.step_calculate(parameters)
     else:
-        d["current_parameters"] = run.current_parameters[run.method]
-        d["chosen_method"] = run.method
+        run.current_step.plot(parameters)
 
-    return JsonResponse(d)
-
-
-def outputs_of_step(request, run_name):
-    """
-    Returns the output keys of a named step of the run. This is used to determine which
-    parameters can be used as input for future steps.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: a JSON response with the output keys
-    :rtype: JsonResponse
-    """
-    run = active_runs[run_name]
-    step_name = request.POST["step_name"]
-    return JsonResponse(run.history.output_keys_of_named_step(step_name), safe=False)
-
-
-def download_plots(request, run_name):
-    """
-    Downloads all plots of the current method in the run. If multiple plots are created,
-    they are zipped together. The format of the plots is specified in the request.
-
-    :param request: the request object
-    :type request: HttpRequest
-    :param run_name: the name of the run
-    :type run_name: str
-
-    :return: a FileResponse with the plots
-    :rtype: FileResponse
-    """
-    run = active_runs[run_name]
-    format_ = request.GET["format"]
-    exported = run.current_plots.export(format_=format_)
-    if len(exported) == 1:
-        filename = f"{run.step_index}-{run.section}-{run.step}-{run.method}.{format_}"
-        return FileResponse(exported[0], filename=filename, as_attachment=True)
-
-    with tempfile.NamedTemporaryFile(delete=False) as f:
-        temp_filename = f.name
-    with zipfile.ZipFile(temp_filename, "w") as zf:
-        for i, plot in enumerate(exported):
-            filename = (
-                f"{run.step_index}-{run.section}-{run.step}-{run.method}-{i}.{format_}"
-            )
-            zf.writestr(filename, plot.getvalue())
-    return FileResponse(
-        open(temp_filename, "rb"),
-        filename=f"{run.step_index}-{run.section}-{run.step}-{run.method}-{format_}.zip",
-        as_attachment=True,
-    )
+    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
 
 
 def tables(request, run_name, index, key=None):
     if run_name not in active_runs:
-        active_runs[run_name] = Run.continue_existing(run_name)
+        active_runs[run_name] = Run(run_name)
     run = active_runs[run_name]
 
+    # TODO this will change with the update to df_mode
     # use current output when applicable (not yet in history)
-    if index < len(run.history.steps):
-        history_step = run.history.steps[index]
-        outputs = history_step.outputs
-        section = history_step.section
-        step = history_step.step
-        method = history_step.method
-        name = run.history.step_names[index]
+    if index < len(run.steps.previous_steps):
+        outputs = run.steps.previous_steps[index].output
+        section = run.steps.previous_steps[index].section
+        step = run.steps.previous_steps[index].operation
+        method = run.steps.previous_steps[index].display_name
     else:
-        outputs = run.current_out
-        section = run.section
-        step = run.step
-        method = run.method
-        name = None
+        outputs = run.current_outputs
+        section = run.current_step.section
+        step = run.current_step.operation
+        method = run.current_step.display_name
 
     options = []
-    for k, value in outputs.items():
+    for k, value in outputs:
         if isinstance(value, pd.DataFrame) and k != key:
             options.append(k)
 
@@ -777,18 +336,136 @@ def tables(request, run_name, index, key=None):
             section=section,
             step=step,
             method=method,
-            name=name,
             clean_ids="clean-ids" if "clean-ids" in request.GET else "",
         ),
     )
 
 
+def add(request: HttpRequest, run_name: str):
+    """
+    Adds a new method to the run. The method is added as the next step.
+
+    :param request: the request object
+    :type request: HttpRequest
+    :param run_name: the name of the run
+    :type run_name: str
+
+    :return: the rendered detail page of the run, new method visible in sidebar
+    :rtype: HttpResponse
+    """
+    run = active_runs[run_name]
+    method = dict(request.POST)["method"][0]
+
+    step = StepFactory.create_step(method, run.steps)
+    run.step_add(step)
+    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
+
+
+def export_workflow(request: HttpRequest, run_name: str):
+    """
+    Exports the workflow of the run as a JSON file.
+
+    :param request: the request object
+    :type request: HttpRequest
+    :param run_name: the name of the run
+    :type run_name: str
+
+    :return: the rendered detail page of the run
+    :rtype: HttpResponse
+    """
+    run = active_runs[run_name]
+    requested_workflow_name = request.POST["name"]
+    run._workflow_export(requested_workflow_name)
+    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
+
+
+def download_plots(request: HttpRequest, run_name: str):
+    """
+    Downloads all plots of the current method in the run. If multiple plots are created,
+    they are zipped together. The format of the plots is specified in the request.
+
+    :param request: the request object
+    :type request: HttpRequest
+    :param run_name: the name of the run
+    :type run_name: str
+
+    :return: a FileResponse with the plots
+    :rtype: FileResponse
+    """
+
+    run = active_runs[run_name]
+    format_ = request.GET["format"]
+    index = run.steps.current_step_index
+    section = run.current_step.section
+    operation = run.current_step.operation
+    exported = run.current_plots.export(format_=format_)
+    if len(exported) == 1:
+        filename = f"{index}-{section}-{operation}.{format_}"
+        return FileResponse(exported[0], filename=filename, as_attachment=True)
+
+    with tempfile.NamedTemporaryFile(delete=False) as f:
+        temp_filename = f.name
+    with zipfile.ZipFile(temp_filename, "w") as zf:
+        for i, plot in enumerate(exported):
+            filename = f"{index}-{section}-{operation}-{i}.{format_}"
+            zf.writestr(filename, plot.getvalue())
+    return FileResponse(
+        open(temp_filename, "rb"),
+        filename=f"{index}-{section}-{operation}.zip",
+        as_attachment=True,
+    )
+
+
+def delete_step(request: HttpRequest, run_name: str):
+    """
+    Deletes a step/method from the run.
+
+    :param request: the request object
+    :type request: HttpRequest
+    :param run_name: the name of the run
+    :type run_name: str
+
+    :return: the rendered detail page of the run, deleted method no longer visible in sidebar
+    :rtype: HttpResponse
+    """
+    run = active_runs[run_name]
+
+    post = dict(request.POST)
+    index = int(post["index"][0])
+    section = post["section"][0]
+
+    run.step_remove(step_index=index, section=section)
+    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
+
+
+def navigate(request, run_name: str):
+    """
+    Navigates to a specific step/method of the run.
+
+    :param request: the request object
+    :param run_name: the name of the run
+
+    :return: the rendered detail page of the run with the specified step/method
+    """
+    run = active_runs[run_name]
+
+    post = dict(request.POST)
+    index = int(post["index"][0])
+    section_name = post["section_name"][
+        0
+    ]  # TODO can this be done without the section_name, like with the delete_step method?
+
+    run.step_goto(index, section_name)
+    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
+
+
 def tables_content(request, run_name, index, key):
     run = active_runs[run_name]
-    if index < len(run.history.steps):
-        outputs = run.history.steps[index].outputs[key]
+    # TODO this will change with df_mode implementation
+    if index < len(run.steps.previous_steps):
+        outputs = run.steps.previous_steps[index].output[key]
     else:
-        outputs = run.current_out[key]
+        outputs = run.current_outputs[key]
     out = outputs.replace(np.nan, None)
 
     if "clean-ids" in request.GET:
@@ -804,13 +481,35 @@ def tables_content(request, run_name, index, key):
     )
 
 
+def change_method(request, run_name):
+    """
+    Changes the method during a step of a run.
+    This is called when the user selects a new method in the first dropdown of a step.
+
+    :param request: the request object
+    :type request: HttpRequest
+    :param run_name: the name of the run
+    :type run_name: str
+
+    :return: a JSON response object containing the new fields for the selected method
+    :rtype: JsonResponse
+    """
+
+    if run_name not in active_runs:
+        active_runs[run_name] = Run(run_name)
+    run = active_runs[run_name]
+    chosen_method = request.POST["chosen_method"]
+    run.step_change_method(chosen_method)
+
+    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
+
+
 def protein_graph(request, run_name, index: int):
+    if run_name not in active_runs:
+        active_runs[run_name] = Run(run_name)
     run = active_runs[run_name]
 
-    if index < len(run.history.steps):
-        outputs = run.history.steps[index].outputs
-    else:
-        outputs = run.current_out
+    outputs = run.current_outputs
 
     if "graph_path" not in outputs:
         return HttpResponseBadRequest(
@@ -872,3 +571,60 @@ def protein_graph(request, run_name, index: int):
             "used_memory": get_memory_usage(),
         },
     )
+
+
+def fill_form(request: HttpRequest, run_name: str):
+    """
+    Fills the form of the current step with the correct values.
+
+    :param request: the request object
+    :type request: HttpRequest
+    :param run_name: the name of the run
+    :type run_name: str
+
+    :return: the filled form
+    :rtype: HttpResponse
+    """
+    run = active_runs[run_name]
+    method_form = get_filled_form_by_request(request, run)
+    form_html = ""
+    for field in method_form:
+        form_html += f"<div>{field.label_tag()} {field}</div>"
+
+    return HttpResponse(form_html)
+
+
+def add_name(request, run_name):
+    """
+    Adds a name to the results of a calculated method of the run. The name can be used
+    to identify the result and use them later.
+
+    :param request: the request object
+    :type request: HttpRequest
+    :param run_name: the name of the run
+    :type run_name: str
+
+    :return: the rendered detail page of the run
+    :rtype: HttpResponse
+    """
+    run = active_runs[run_name]
+    run.name_step(int(request.POST["index"]), request.POST["name"])
+    return HttpResponseRedirect(reverse("runs:detail", args=(run_name,)))
+
+
+def download_table(request, run_name, index, key):
+    if run_name not in active_runs:
+        active_runs[run_name] = Run(run_name)
+    run = active_runs[run_name]
+
+    instance_id = run.steps.all_steps[index].instance_identifier
+    buffer = io.StringIO()
+    df: pd.DataFrame = run.steps.get_step_output(
+        Step, key, instance_id, include_current_step=True
+    )
+    df.to_csv(buffer)
+
+    buffer.seek(0)
+    csv_bytes = buffer.getvalue()
+
+    return FileResponse(csv_bytes, content_type="text/csv")
