--- ui/runs_v2/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class RunsV2Config(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "runs_v2"
