--- ui/runs_v2/static/runs_v2/index.css
@@ -1,8 +1,0 @@
-#contain {
-    height:100vh;
-    margin-top:-3.6rem;
-}
-
-#row {
-    width:80%;
-}
