--- ui/runs_v2/static/runs_v2/sidebar.css
@@ -1,54 +1,0 @@
-p {
-    margin: 5px;
-}
-
-.selected {
-    background-color: #E8EDF3;
-    border-radius: 3px;
-}
-
-.finished {
-    color: #adadad;
-    border-radius: 15px;
-}
-
-.full_width {
-    width: 100%;
-}
-
-.simple-tags > ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    display: -webkit-box;
-    display: -webkit-flex;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-flex-wrap: wrap;
-    -ms-flex-wrap: wrap;
-    flex-wrap: wrap
-}
-
-.simple-tags ul li {
-    padding: 3px;
-}
-
-#explainAddStep {
-    font-size: small;
-}
-
-.table tr:last-child td{
-    border:none;
-}
-
-.table {
-    margin-bottom: -0.05rem;
-}
-
-.removeBtn {
-    padding: 0px;
-}
-
-.removeIcon:hover {
-    fill: #444445;
-}
