--- ui/runs_v2/templates/runs_v2/field_name_output_text.html
@@ -1,11 +1,0 @@
-{# params: key, default, disabled, form #}
-
-{# key: the name of the input #}
-{# default: the default value of the input #}
-{# disabled: if the input is disabled #}
-{# form: the name of the form #}
-
-<input name="{{ key }}" type="text" id="{{ key }}" value="{{ default }}"
-    class="form-control" {% if disabled %} disabled {% endif %}
-    {%if form %} form="{{form}}" {% endif %}
-    placeholder="name output">
