--- ui/runs_v2/templates/runs_v2/field_named.html
@@ -1,16 +1,0 @@
-{# params: key name disabled default steps outputs #}
-{% if class %}
-    <div class="{{ class }}" id="{{ key }}_div"> {% endif %}
-<label for="{{ key }}">{{ name }}</label>
-<select name="{{ key }}" id="{{ key }}" class="named form-select" {% if disabled %} disabled {% endif %} >
-    {% for option in steps %}
-        <option value="{{ option }}" {% if option == default.0 %} selected {% endif %} >{{ option }}</option>
-    {% endfor %}
-</select>
-<select name="{{ key }}" class="form-select"{% if not disabled %} id="{{ key }}_outputs" {% endif %} {% if disabled %}
-        disabled {% endif %} >
-    {% for option in outputs %}
-        <option value="{{ option }}" {% if option == default.1 %} selected {% endif %} >{{ option }}</option>
-    {% endfor %}
-</select>
-{% if class %}</div>{% endif %}
