--- ui/runs_v2/templates/runs_v2/field_named_with_fields.html
@@ -1,4 +1,0 @@
-{% include "runs/field_named.html" with key=key name=name steps=steps disabled=disabled outputs=outputs %}
-{% for field in additional_fields %}
-    {{field|safe}}
-{% endfor %}
