--- ui/runs_v2/templates/runs_v2/form_export_workflow.html
@@ -1,8 +1,0 @@
-{# params: default #}
-<form action="{% url 'runs_v2:export_workflow' run_name %}" method="post">
-    {% csrf_token %}
-    <div class="input-group">
-        <input type="text" name="name" value="{{ default }}" class="form-control">
-        <input type="submit" value="export workflow" class="btn btn-outline-blue">
-    </div>
-</form>
