--- ui/runs_v2/templates/runs_v2/sidebar.html
@@ -1,49 +1,0 @@
-{% load static %}
-{# params: workflow_steps run_name#}
-{% block css %}
-    <link rel="stylesheet" type="text/css" href="{% static 'runs_v2/sidebar.css' %}">
-{% endblock %}
-
-<div class="container container-fixed-width" style="width: 2000px; max-width: 100%;">
-    <div id="accordion" class="">
-        {% for section in workflow_steps %}
-            <div class="border-top">
-                <div class="card-header d-flex btn p-3 align-items-center bg-white border-bottom" data-bs-toggle="collapse"
-                     href="#collapse_{{ section.id }}">
-                    
-                    {# add section icon #}
-                    {% if section.id == 'importing' %}
-                        <img src="{% static 'img/download-file-icon.svg' %}" alt="toggle sidebar" width="20" height="20" class="me-1">
-                    {% elif section.id == 'data_preprocessing' %}
-                        <svg xmlns="http://www.w3.org/2000/svg" width="23" height="23" fill="#4A536A" class="bi bi-gear-fill" viewBox="0 0 16 16">
-                            <path d="M9.405 1.05c-.413-1.4-2.397-1.4-2.81 0l-.1.34a1.464 1.464 0 0 1-2.105.872l-.31-.17c-1.283-.698-2.686.705-1.987 1.987l.169.311c.446.82.023 1.841-.872 2.105l-.34.1c-1.4.413-1.4 2.397 0 2.81l.34.1a1.464 1.464 0 0 1 .872 2.105l-.17.31c-.698 1.283.705 2.686 1.987 1.987l.311-.169a1.464 1.464 0 0 1 2.105.872l.1.34c.413 1.4 2.397 1.4 2.81 0l.1-.34a1.464 1.464 0 0 1 2.105-.872l.31.17c1.283.698 2.686-.705 1.987-1.987l-.169-.311a1.464 1.464 0 0 1 .872-2.105l.34-.1c1.4-.413 1.4-2.397 0-2.81l-.34-.1a1.464 1.464 0 0 1-.872-2.105l.17-.31c.698-1.283-.705-2.686-1.987-1.987l-.311.169a1.464 1.464 0 0 1-2.105-.872l-.1-.34zM8 10.93a2.929 2.929 0 1 1 0-5.86 2.929 2.929 0 0 1 0 5.858z"/>
-                        </svg>
-                    {% elif section.id == 'data_analysis' %}
-                        <svg xmlns="http://www.w3.org/2000/svg" width="23" height="23" fill="#4A536A" class="bi bi-bar-chart-fill" viewBox="0 0 16 16">
-                            <path d="M1 11a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1v3a1 1 0 0 1-1 1H2a1 1 0 0 1-1-1v-3zm5-4a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1v7a1 1 0 0 1-1 1H7a1 1 0 0 1-1-1V7zm5-5a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1v12a1 1 0 0 1-1 1h-2a1 1 0 0 1-1-1V2z"/>
-                        </svg>
-                    {% elif section.id == 'data_integration' %}
-                        <svg xmlns="http://www.w3.org/2000/svg" width="26" height="27" fill="#4A536A"  class="bi bi-database-fill-add" viewBox="0 0 16 16" style="margin-left:-0.125rem;">
-                            <path d="M12.5 16a3.5 3.5 0 1 0 0-7 3.5 3.5 0 0 0 0 7Zm.5-5v1h1a.5.5 0 0 1 0 1h-1v1a.5.5 0 0 1-1 0v-1h-1a.5.5 0 0 1 0-1h1v-1a.5.5 0 0 1 1 0ZM8 1c-1.573 0-3.022.289-4.096.777C2.875 2.245 2 2.993 2 4s.875 1.755 1.904 2.223C4.978 6.711 6.427 7 8 7s3.022-.289 4.096-.777C13.125 5.755 14 5.007 14 4s-.875-1.755-1.904-2.223C11.022 1.289 9.573 1 8 1Z"/>
-                            <path d="M2 7v-.839c.457.432 1.004.751 1.49.972C4.722 7.693 6.318 8 8 8s3.278-.307 4.51-.867c.486-.22 1.033-.54 1.49-.972V7c0 .424-.155.802-.411 1.133a4.51 4.51 0 0 0-4.815 1.843A12.31 12.31 0 0 1 8 10c-1.573 0-3.022-.289-4.096-.777C2.875 8.755 2 8.007 2 7Zm6.257 3.998L8 11c-1.682 0-3.278-.307-4.51-.867-.486-.22-1.033-.54-1.49-.972V10c0 1.007.875 1.755 1.904 2.223C4.978 12.711 6.427 13 8 13h.027a4.552 4.552 0 0 1 .23-2.002Zm-.002 3L8 14c-1.682 0-3.278-.307-4.51-.867-.486-.22-1.033-.54-1.49-.972V13c0 1.007.875 1.755 1.904 2.223C4.978 15.711 6.427 16 8 16c.536 0 1.058-.034 1.555-.097a4.507 4.507 0 0 1-1.3-1.905Z"/>
-                          </svg>
-                    {% endif %}
-                    
-                    <div class="ms-3">
-                        {{ section.name }}
-                    </div>
-                </div>
-                <div id="collapse_{{ section.id }}"
-                     class="collapse{% if section.selected %} show {% endif %}">
-                    <div class="card-body my-2">
-                        {% include "runs_v2/sidebar_section.html" with section=section run_name=run_name %}
-                    </div>
-                </div>
-            </div>
-        {% endfor %}
-    </div>
-
-    <div class="mt-4">
-        {% include "runs_v2/form_export_workflow.html" with run_name=run_name default="workflow_"|add:run_name %}
-    </div>
-</div>
