--- ui/runs_v2/templates/runs_v2/sidebar_section.html
@@ -1,82 +1,0 @@
-{# params: run_name section{id, name, finished, selected, possible_steps{id, name, methods{id, name, description}}, steps{id, name, finished, selected, index, method_name}} #}
-
-{% block js %}
-    <script>
-        $(function () {
-            $('[data-toggle="tooltip"]').tooltip({
-                delay: {show: 1000, hide: 0}
-            });
-        });
-    </script>
-{% endblock %}
-
-<div>
-    {% for step in section.steps %}
-        <div class="wrapper {% if step.selected %}selected{% endif %} {% if step.finished %}finished{% endif %} d-flex align-items-center">
-            <form action="{% url 'runs_v2:navigate' run_name %}" method="post">
-                {% csrf_token %}
-                <input type="hidden" name="section_name" value="{{ section.id }}">
-                <input type="hidden" name="index" value="{{ step.index }}">
-                <button class="me-1" style="border: none; background: none; text-align: left">
-                    <p class="me-1">{{ step.name }}: <i>{{ step.method_name }}</i></p>
-                </button>
-            </form>
-
-            {% if not step.selected and not step.finished %}
-                {# button to remove step from workflow #}
-                <form action="{% url 'runs_v2:delete_step' run_name %}" method="post">
-                    {% csrf_token %}
-                    <input type="hidden" name="index" value="{{ step.index }}">
-                    <input type="hidden" name="section" value="{{ step.section }}">
-                    <button class="removeBtn btn d-flex justify-content-center align-items-end mt-1">
-                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="grey" class="bi bi-x removeIcon">
-                        <path d="M4.646 4.646a.5.5 0 0 1 .708 0L8 7.293l2.646-2.647a.5.5 0 0 1 .708.708L8.707 8l2.647 2.646a.5.5 0 0 1-.708.708L8 8.707l-2.646 2.647a.5.5 0 0 1-.708-.708L7.293 8 4.646 5.354a.5.5 0 0 1 0-.708z"/>
-                    </svg></button>
-                </form>
-            {% endif %}
-
-        </div>
-    {% endfor %}
-</div>
-
-{% if section.finished %}
-    <p class="pb-2"><i class="text-muted">Go back to this section to add more steps</i></p>
-{% else %}
-    <button class="btn btn-blue full_width mt-2 mb-2" type="button" data-bs-toggle="collapse"
-            data-bs-target="#add_steps_div_{{ section.id }}">
-        + add steps
-    </button>
-
-    <div id="add_steps_div_{{ section.id }}" class="collapse">
-        <div class="d-flex justify-content-center mb-2">
-            <i id="explainAddStep" class="text-muted">Click on a method to add it to your run</i>
-        </div>
-
-        <table class="table">
-            {% for possible_step in section.possible_steps %}
-                <tr>
-                    <td class="small">{{ possible_step.name }}</td>
-                    <td>
-                        <div class="simple-tags">
-                            <ul>
-                                {% for method in possible_step.methods %}
-                                    <li>
-                                        <form action="{% url 'runs_v2:add' run_name %}" method="post">
-                                            {% csrf_token %}
-                                            <input type="hidden" name="section_name" value="{{ section.id }}">
-                                            <input type="hidden" name="step" value="{{ possible_step.id }}">
-                                            <input type="hidden" name="method" value="{{ method.id }}">
-                                            <button class="btn btn-sm btn-grey" data-toggle="tooltip"
-                                                    data-placement="right"
-                                                    title="{{ method.description }}">{{ method.name }}</button>
-                                        </form>
-                                    </li>
-                                {% endfor %}
-                            </ul>
-                        </div>
-                    </td>
-                </tr>
-            {% endfor %}
-        </table>
-    </div>
-{% endif %}
