--- ui/runs_v2/templates/runs_v2/tables.html
@@ -1,73 +1,0 @@
-{% extends 'base.html' %}
-{% load static %}
-
-{% block css %}
-    <link rel="stylesheet" type="text/css" href="{% static 'css/datatables-1.13.4.min.css' %}">
-    <link rel="stylesheet" type="text/css" href="{% static 'runs/style.css' %}">
-{% endblock %}
-
-{% block js %}
-    <script type="text/javascript" src="{% static 'js/jquery.js' %}"></script>
-    <script type="text/javascript" src="{% static 'js/datatables-1.13.4.min.js' %}"></script>
-    <script>
-        $(document).ready(function () {
-            $.ajax({
-                url: "{% url 'runs_v2:tables_content' run_name index key %}",
-                data: "{{ clean_ids }}",
-                type: "GET",
-                success: function (response) {
-                    let tr = $("<tr>");
-                    response.columns.forEach(element => {
-                        tr.append($("<th>").text(element))
-                    });
-                    $("#datatable").html($("<thead>").append(tr))
-                        .DataTable({
-                            data: response.data,
-                            columnDefs: [{
-                                targets: "_all",
-                                render: function (data, type, row) {
-                                    if (type !== 'display') {
-                                        return data;
-                                    }
-                                    if (typeof data === 'string' && data.startsWith("https://")) {
-                                        return data.split(' ').map(url => `<a href="${url}" target="_blank">${url.split('/').pop()}</a>`).join(' ')
-                                    }
-                                    if (typeof data === 'number' && (!Number.isInteger(data) || data >= 1e4)) {
-                                        return data.toPrecision(4);
-                                    }
-                                    return `<div style="word-break:break-all">${data}</div>`;
-                                },
-                            }],
-                        });
-                }
-            });
-            $('#tables_dropdown').on("change", function () {
-                window.location.href = $(this).val()
-            })
-        });
-    </script>
-{% endblock %}
-
-{% block content %}
-    <div class="p-3">
-        <p>
-            Index {{ index|add:1 }}, Section {{ section }}, step {{ step }}, method {{ method }}
-            <br>
-        </p>
-        <div class="d-flex justify-content-between">
-            <div class="input-group mb-3 w-50">
-                <span class="input-group-text">Choose table</span>
-                {% include 'runs/field_select_with_label.html' with key="tables_dropdown" categories=options only %}
-            </div>
-            <div>
-            {% if clean_ids %}
-                <a class="mb-3 btn btn-grey" href="{% url 'runs:tables' run_name index key %}">Enable Isoforms</a>
-            {% else %}
-                <a class="mb-3 btn btn-grey" href="?clean-ids">Disable Isoforms</a>
-            {% endif %}
-            <a class="mb-3 btn btn-grey" href="{% url 'runs_v2:download_table' run_name index key %}">Download Table</a>
-            </div>
-        </div>
-        <table id="datatable" class="display"></table>
-    </div>
-{% endblock %}
